requirements.txt
nbformat>=5.7.0
typing-extensions>=4.4.0
black>=23.12.0
openai>=1.6.0
python-dotenv>=1.0.0
